refactor(auth): clean up paypal files into its own module
Because:

* We want to extend the types/errors used by paypal in a more
  organized manner.

This commit:

* Refactors the paypal library code into its own module.

Issue #11592
