+++ packages/fxa-auth-server/bin/key_server.js
@@ -84,8 +84,8 @@ async function run(config) {
     Container.set(StripeHelper, stripeHelper);
 
     if (config.subscriptions.paypalNvpSigCredentials.enabled) {
-      const { PayPalClient } = require('../lib/payments/paypal-client');
-      const { PayPalHelper } = require('../lib/payments/paypal');
+      const { PayPalClient } = require('../lib/payments/paypal/client');
+      const { PayPalHelper } = require('../lib/payments/paypal/helper');
       const paypalClient = new PayPalClient(
         config.subscriptions.paypalNvpSigCredentials
       );
