+++ packages/fxa-auth-server/lib/payments/paypal/client.ts
@@ -1,10 +1,10 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+import { EventEmitter } from 'events';
 import pRetry from 'p-retry';
 import querystring from 'querystring';
 import superagent from 'superagent';
-import { EventEmitter } from 'events';
 
 export const PAYPAL_SANDBOX_BASE = 'https://api-3t.sandbox.paypal.com';
 export const PAYPAL_SANDBOX_IPN_BASE = 'https://ipnpb.sandbox.paypal.com';
