+++ packages/fxa-auth-server/lib/routes/subscriptions/paypal.ts
@@ -18,10 +18,12 @@ import Container from 'typedi';
 
 import { ConfigType } from '../../../config';
 import error from '../../error';
-import { COUNTRIES_LONG_NAME_TO_SHORT_NAME_MAP } from '../../payments/stripe';
-import { PayPalHelper } from '../../payments/paypal';
+import { PayPalHelper } from '../../payments/paypal/helper';
 import STATES_LONG_NAME_TO_SHORT_NAME_MAP from '../../payments/states-long-name-to-short-name-map.json';
-import { StripeHelper } from '../../payments/stripe';
+import {
+  COUNTRIES_LONG_NAME_TO_SHORT_NAME_MAP,
+  StripeHelper,
+} from '../../payments/stripe';
 import { reportSentryError } from '../../sentry';
 import { msToSec } from '../../time';
 import { AuthLogger, AuthRequest } from '../../types';
