+++ packages/fxa-auth-server/scripts/paypal-processor.ts
@@ -6,9 +6,9 @@ import { StatsD } from 'hot-shots';
 import Container from 'typedi';
 import { promisify } from 'util';
 
-import { PayPalHelper } from '../lib/payments/paypal';
-import { PayPalClient } from '../lib/payments/paypal-client';
-import { PaypalProcessor } from '../lib/payments/paypal-processor';
+import { PayPalHelper } from '../lib/payments/paypal/helper';
+import { PayPalClient } from '../lib/payments/paypal/client';
+import { PaypalProcessor } from '../lib/payments/paypal/processor';
 import { setupProcessingTaskObjects } from '../lib/payments/processing-tasks-setup';
 
 const pckg = require('../package.json');
