+++ packages/fxa-auth-server/test/local/routes/account.js
@@ -18,7 +18,7 @@ const log = require('../../../lib/log');
 const otplib = require('otplib');
 const { default: Container } = require('typedi');
 const { StripeHelper } = require('../../../lib/payments/stripe');
-const { PayPalHelper } = require('../../../lib/payments/paypal');
+const { PayPalHelper } = require('../../../lib/payments/paypal/helper');
 const { CapabilityService } = require('../../../lib/payments/capability');
 const { normalizeEmail } = require('fxa-shared').email.helpers;
 const { MozillaSubscriptionTypes } = require('fxa-shared/subscriptions/types');
