+++ packages/fxa-auth-server/test/local/routes/subscriptions/paypal.js
@@ -12,7 +12,7 @@ const { filterCustomer } = require('fxa-shared/subscriptions/stripe');
 const error = require('../../../../lib/error');
 const { getRoute } = require('../../../routes_helpers');
 const mocks = require('../../../mocks');
-const { PayPalHelper } = require('../../../../lib/payments/paypal');
+const { PayPalHelper } = require('../../../../lib/payments/paypal/helper');
 const uuid = require('uuid');
 const { StripeHelper } = require('../../../../lib/payments/stripe');
 const customerFixture = require('../../payments/fixtures/stripe/customer1.json');
