+++ packages/fxa-auth-server/test/local/routes/subscriptions/stripe-webhooks.js
@@ -44,7 +44,7 @@ const eventTaxRateCreated = require('../../payments/fixtures/stripe/event_tax_ra
 const eventTaxRateUpdated = require('../../payments/fixtures/stripe/event_tax_rate_created.json');
 const eventPaymentMethodUpdated = require('../../payments/fixtures/stripe/event_payment_method_updated.json');
 const { default: Container } = require('typedi');
-const { PayPalHelper } = require('../../../../lib/payments/paypal');
+const { PayPalHelper } = require('../../../../lib/payments/paypal/helper');
 const { CapabilityService } = require('../../../../lib/payments/capability');
 const { CurrencyHelper } = require('../../../../lib/payments/currencies');
 const { asyncIterable } = require('../../../mocks');
